+++ packages/ai/src/agent/do-stream-step.ts
@@ -21,7 +21,10 @@ export async function doStreamStep(
   conversationPrompt: LanguageModelV2Prompt,
   modelInit: string | (() => Promise<LanguageModelV2>),
   writable: WritableStream<UIMessageChunk>,
-  tools?: LanguageModelV2CallOptions['tools']
+  tools?: LanguageModelV2CallOptions['tools'],
+  options?: {
+    sendStart?: boolean;
+  }
 ) {
   'use step';
 
@@ -65,9 +68,11 @@ export async function doStreamStep(
     .pipeThrough(
       new TransformStream<LanguageModelV2StreamPart, UIMessageChunk>({
         start: (controller) => {
-          controller.enqueue({
-            type: 'start',
-          });
+          if (options?.sendStart) {
+            controller.enqueue({
+              type: 'start',
+            });
+          }
           controller.enqueue({
             type: 'start-step',
           });
@@ -75,9 +80,6 @@ export async function doStreamStep(
         flush: (controller) => {
           controller.enqueue({
             type: 'finish-step',
-          });
-          controller.enqueue({
-            type: 'finish',
           });
         },
         transform: async (part, controller) => {